fix no-use-before-declaration issues
