+++ org.eclipse.winery.repository.ui/src/app/instance/imports/listDefinedTypesAndElements.component.ts
@@ -73,8 +73,3 @@ export class ListDefinedTypesAndElementsComponent implements OnInit {
         this.notify.error(error.message);
     }
 }
-
-class SingleColumn {
-    constructor(public readonly key: string) {
-    }
-}
