+++ org.eclipse.winery.repository.ui/src/app/section/entityContainer/differencesData.ts
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 Contributors to the Eclipse Foundation
+ * Copyright (c) 2018 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -11,25 +11,9 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-export class TargetInterfaceOperation {
-    interfaceName = '';
-    operationName = '';
-}
+import { SectionData } from '../sectionData';
 
-export class NodeOperation extends TargetInterfaceOperation {
-    nodeRef = '';
-}
-
-export class RelationshipOperation extends TargetInterfaceOperation {
-    relationshipRef = '';
-}
-
-export class PlanOperation {
-    planRef = '';
-}
-
-export enum CurrentSelectedEnum {
-    nodeTemplate = 'nodetypes',
-    relationshipTemplate = 'relationshiptypes',
-    plan = 'plan'
+export class DifferencesData {
+    base: SectionData;
+    diff: string;
 }
