+++ org.eclipse.winery.repository.ui/src/app/winerySpinnerWithInfinityModule/winerySpinnerWithInfinity.component.ts
@@ -11,16 +11,10 @@
  *
  * SPDX-License-Identifier: EPL-2.0 OR Apache-2.0
  *******************************************************************************/
-import {Component, forwardRef, Input} from '@angular/core';
-import {ControlValueAccessor, NG_VALUE_ACCESSOR} from '@angular/forms';
+import { Component, forwardRef, Input } from '@angular/core';
+import { ControlValueAccessor, NG_VALUE_ACCESSOR } from '@angular/forms';
 
 const noop = () => {
-};
-
-const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
-    provide: NG_VALUE_ACCESSOR,
-    useExisting: forwardRef(() => SpinnerWithInfinityComponent),
-    multi: true
 };
 
 /**
@@ -46,7 +40,13 @@ const CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR: any = {
 @Component({
     selector: 'winery-spinner-with-infinity',
     templateUrl: 'winerySpinnerWithInfinity.component.html',
-    providers: [CUSTOM_INPUT_CONTROL_VALUE_ACCESSOR]
+    providers: [
+        {
+            provide: NG_VALUE_ACCESSOR,
+            useExisting: forwardRef(() => SpinnerWithInfinityComponent),
+            multi: true
+        }
+    ]
 })
 
 export class SpinnerWithInfinityComponent implements ControlValueAccessor {
